* new upload dialog * added import dialog * user has more control over import process now * lots of refactoring

+++ src/main/java/com/dude/dms/brain/FileManager.kt
@@ -2,7 +2,6 @@ package com.dude.dms.brain
 
 import com.dude.dms.brain.DmsLogger.Companion.getLogger
 import com.dude.dms.brain.options.Options
-import org.apache.commons.net.ftp.FTPClient
 import org.apache.pdfbox.pdmodel.PDDocument
 import org.apache.pdfbox.pdmodel.PDPage
 import org.apache.pdfbox.pdmodel.PDPageContentStream
@@ -23,30 +22,16 @@ import javax.imageio.ImageIO
 @Component
 class FileManager {
 
-    fun testFtp(): Boolean {
-        val ftpClient = FTPClient()
-        return try {
-            try {
-                ftpClient.connectTimeout = 3000
-                ftpClient.connect(Options.get().storage.ftp.url, Options.get().storage.ftp.port)
-                ftpClient.login(Options.get().storage.ftp.url, Options.get().storage.ftp.password)
-                ftpClient.listNames()
-                ftpClient.logout()
-                true
-            } finally {
-                ftpClient.disconnect()
-            }
-        } catch (e: IOException) {
-            LOGGER.error(e.message!!)
-            false
-        }
-    }
-
     fun getPdf(guid: String) = File("${Options.get().doc.savePath}/pdf/$guid.pdf")
 
     fun getFirstImage(guid: String) = File("${Options.get().doc.savePath}/img/${guid}_00.png")
 
-    fun getImages(guid: String) = File("${Options.get().doc.savePath}/img/").listFiles { _, name -> name.startsWith(guid) }
+    fun getImages(guid: String) = File("${Options.get().doc.savePath}/img/").listFiles { _, name -> name.startsWith(guid) }!!
+
+    fun delete(guid: String) {
+        getPdf(guid).delete()
+        getImages(guid).forEach { it.delete() }
+    }
 
     fun createDirectories() {
         val saveDir = File(Options.get().doc.savePath)
@@ -58,12 +43,15 @@ class FileManager {
         }
     }
 
-    fun importPdf(pdf: File): String? {
+    fun importPdf(pdf: File, move: Boolean = true): String? {
         val guid = UUID.randomUUID().toString()
         val targetPath = Paths.get(Options.get().doc.savePath, "pdf", "$guid.pdf")
         LOGGER.info(t("pdf.import", pdf.name))
         try {
-            Files.move(pdf.toPath(), targetPath, StandardCopyOption.ATOMIC_MOVE).toFile()
+            when {
+                move -> Files.move(pdf.toPath(), targetPath, StandardCopyOption.ATOMIC_MOVE).toFile()
+                else -> Files.copy(pdf.toPath(), targetPath, StandardCopyOption.REPLACE_EXISTING).toFile()
+            }
         } catch (e: IOException) {
             LOGGER.error(e.message!!, e)
             return null
@@ -73,12 +61,12 @@ class FileManager {
         return guid
     }
 
-    fun importImage(img: File): String? {
+    fun importImage(img: File, move: Boolean = true): String? {
         val guid = UUID.randomUUID().toString()
         LOGGER.info(t("image.import", img.name))
         try {
             ImageIO.write(processImg(ImageIO.read(img)), "png", Paths.get(Options.get().doc.savePath, "img", "${guid}_00.png").toFile())
-            img.delete()
+            if (move) img.delete()
         } catch (e: IOException) {
             LOGGER.error(e.message!!, e)
             return null
